+++ src/shared.js
@@ -23,3 +23,14 @@ export function filterUndefined(obj) {
     return filtered;
   }, {});
 }
+
+export function getOptionsAndCallback(optionsOrCallback, callback) {
+  let options;
+  if (typeof optionsOrCallback === 'function') {
+    callback = optionsOrCallback;
+  } else {
+    options = optionsOrCallback;
+  }
+
+  return [options, callback];
+}